Extracted most of CreateAccount.jsx into a component
